+++ solana/rpc/async_api.py
@@ -1120,7 +1120,7 @@ class AsyncClient(_ClientCore):  # pylint: disable=too-many-public-methods
         return await self.confirm_transaction(resp["result"], conf_comm)
 
     async def confirm_transaction(
-        self, tx_sig: str, commitment: Commitment = Finalized, sleep_seconds: int = 0.5
+        self, tx_sig: str, commitment: Commitment = Finalized, sleep_seconds: float = 0.5
     ) -> types.RPCResponse:
         """Confirm the transaction identified by the specified signature.
